backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
openpyxl==3.1.2
mysql-connector-python==8.4.0
